Extract service layer from commands
Create 6 focused service classes from inline data logic in commands:

- PackageResolverService: version discovery, cache lookup, download, extraction
  (unified from PackageCommand + AssemblyCommand duplicates)
- LibraryMetadataService: assembly inspection, PDB audit, SourceLink verification,
  builder inference, transitive reference resolution (~340 lines from AssemblyCommand)
- TypeSearchService: unified 6-source type collection with optional pattern filtering
  (merged two near-duplicate 200-line paths from FindCommand)
- ProjectAssetsParser: project.assets.json parsing (from FindCommand)
- DependencyResolutionService: transitive NuGet dependency tree resolution
  (from PackageCommand)
- PackageMetadataService: NuGet metadata enrichment — published date, downloads,
  deprecation, vulnerabilities, package size (~400 lines from PackageCommand)

Commands now delegate to services for data logic, keeping only CLI dispatch,
argument parsing, and output formatting. Net reduction of 235 lines with
elimination of duplicate GetLatestVersionAsync, cache/download, and type
collection code.
